git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@8914 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/descent.h
@@ -2496,7 +2496,7 @@ class CAutoNetGame {
 		char					szName [81];		//game name
 		int					nLevel;
 		ubyte					ipAddr [4];
-		ushort				nPort;
+		int					nPort;
 		ubyte					uConnect;
 		ubyte					uType;
 		ubyte					bHost;
